Move Admin.Measure to admin package and rename to AdminMeasure
